requirements.txt
opencv-python==4.7.0.72
numpy==1.24.1
Pillow==9.4.0

